refactor: move dto to dtos class
