+++ src/main/kotlin/com/technews/aggregate/releases/dto/ReleaseDtos.kt
@@ -8,6 +8,27 @@ import java.util.*
 
 private val logger = KotlinLogging.logger {}
 
+data class ReleaseResponse(
+    val project: String,
+    val version: String,
+    val date: String,
+    val url: String,
+    val tags: List<String>,
+    val createdDt: String,
+) {
+    companion object {
+        fun from(release: Release): ReleaseResponse =
+            ReleaseResponse(
+                project = release.project,
+                version = release.version,
+                date = release.date,
+                url = release.url,
+                tags = release.tags,
+                createdDt = release.createdDt,
+            )
+    }
+}
+
 data class SaveReleaseRequest(
     val project: String = "",
     val version: String = "",
