--- src/main/kotlin/com/technews/aggregate/releases/dto/ReleaseResponse.kt
@@ -1,24 +1,0 @@
-package com.technews.aggregate.releases.dto
-
-import com.technews.aggregate.releases.domain.Release
-
-data class ReleaseResponse(
-    val project: String,
-    val version: String,
-    val date: String,
-    val url: String,
-    val tags: List<String>,
-    val createdDt: String,
-) {
-    companion object {
-        fun from(release: Release): ReleaseResponse =
-            ReleaseResponse(
-                project = release.project,
-                version = release.version,
-                date = release.date,
-                url = release.url,
-                tags = release.tags,
-                createdDt = release.createdDt,
-            )
-    }
-}
